backend/requirements.txt
autopep8==1.4.3
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
pep8==1.7.1
Pillow==5.4.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pytz==2018.9
